+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 # 12.0.0-alpha.6 (Unreleased)
 - Fix exponential notation syntax. https://github.com/rescript-lang/rescript/pull/7174
 
+
 #### :bug: Bug fix
 - Fix bug where a ref assignment is moved ouside a conditional. https://github.com/rescript-lang/rescript/pull/7176
 - Fix nullable to opt conversion. https://github.com/rescript-lang/rescript/pull/7193
@@ -24,6 +25,10 @@
 - Ast cleanup: explicit representation for optional record fields in types. https://github.com/rescript-lang/rescript/pull/7190 https://github.com/rescript-lang/rescript/pull/7191
 - AST cleanup: first-class expression and patterns for records with optional fields. https://github.com/rescript-lang/rescript/pull/7192
 
+
+#### :house: Internal
+
+- Add attribute `res.array{Get,Set}` for `Array.{get, set}`. https://github.com/rescript-lang/rescript-compiler/pull/6343
 
 # 12.0.0-alpha.5